Added import directory feature and fixed some bugs

+++ DLLirant.NET/Classes/FileOperations.cs
@@ -32,9 +32,22 @@ namespace DLLirant.NET.Classes
 
         public void RenameFile(string path, string newpath)
         {
-            if (File.Exists(path))
+            if (File.Exists(path) && !File.Exists(newpath))
             {
                 File.Move(path, newpath);
+            }
+        }
+
+        public void CopyFilesDirToDir(string dllname, string sourceDir, string targetDir)
+        {
+            if (Directory.Exists(sourceDir))
+            {
+                foreach (string file in Directory.GetFiles(sourceDir)) {
+                    if (Path.GetFileName(file) != dllname)
+                    {
+                        File.Copy(file, Path.Combine(targetDir, Path.GetFileName(file)));
+                    }
+                }
             }
         }
     }
